s390x: auto generate

+++ builtin.go
@@ -61,7 +61,7 @@ func X__builtin_ctz(t *TLS, n uint32) int32 {
 }
 
 func X__builtin_ctzl(tls *TLS, x ulong) int32 {
-	return int32(mbits.TrailingZeros64(uint64(x)))
+	return int32(mbits.TrailingZeros64(x))
 }
 
 func X__builtin_clz(t *TLS, n uint32) int32 {
@@ -69,7 +69,7 @@ func X__builtin_clz(t *TLS, n uint32) int32 {
 }
 
 func X__builtin_clzl(t *TLS, n ulong) int32 {
-	return int32(mbits.LeadingZeros64(uint64(n)))
+	return int32(mbits.LeadingZeros64(n))
 }
 
 func X__builtin_clzll(t *TLS, n uint64) int32 {
@@ -174,7 +174,7 @@ func X__builtin_popcount(t *TLS, x uint32) int32 {
 
 // int __builtin_popcountl (unsigned long x)
 func X__builtin_popcountl(t *TLS, x ulong) int32 {
-	return int32(mbits.OnesCount64(uint64(x)))
+	return int32(mbits.OnesCount64(x))
 }
 
 // char * __builtin___strcpy_chk (char *dest, const char *src, size_t os);
